IBM|𝘼𝙄𝙒𝙖𝙩𝙨𝙤𝙣𝙇𝙖𝙗 💻🎓/ibmlearner_conversationalAI/server.py
from dotenv import load_dotenv
from flask import Flask, render_template, request, Response
from ibmservices import ibmservices
import os

app = Flask(__name__)
response_text = None

@app.route('/')
def file_uploader():
   return render_template('upload.html')

@app.route('/audio/<filename>')
def stream_mp3(filename):
    def generate():
        with open(filename, 'rb') as fmp3:
            data = fmp3.read(1024)
            while data:
                yield data
                data = fmp3.read(1024)
    return Response(generate(), mimetype="audio/mpeg3")

@app.route('/uploader', methods = ['POST'])
def upload_file():
   if request.method == 'POST':
        f = request.files['file']
        try:
            if f.filename != '':
                l = len(f.filename)
                extn = f.filename[l-3:l]
                if extn not in ["mp3","wav"]:
                    raise Exception("Sorry, the file type is unsupported. Try .mp3 or .wav files")
                f.save(f.filename)
                stt_text=ibmservices.speechToText(f.filename,extn)
                os.remove(f.filename)
                return ibmservices.getResponseFromAssistant(stt_text)
            else:
                raise Exception("Sorry. No filename recognized")
        except Exception as excp:
            print(excp.__traceback__)
            return str(excp),500

if __name__ == "__main__":
    app.run(host="0.0.0.0", port=8080)