IBM|𝘼𝙄𝙒𝙖𝙩𝙨𝙤𝙣𝙇𝙖𝙗 💻🎓/ibmlearner_conversationalAI/requirements.txt
PyJWT==1.7.1 
flask-appbuilder==3.4.4 
markupsafe==1.1.1 
httpx==0.20 
Flask 
ibm-watson 
python-dotenv 
Flask-Cors
